Simplify symbol processor by breaking code out into other packages
