+++ mockingbird/processor/src/main/kotlin/com/anthonycr/mockingbird/processor/MockingbirdSymbolProcessorProvider.kt
@@ -1,5 +1,7 @@
 package com.anthonycr.mockingbird.processor
 
+import com.anthonycr.mockingbird.processor.internal.generator.FakeFunctionGenerator
+import com.anthonycr.mockingbird.processor.internal.generator.FakeImplementationGenerator
 import com.google.devtools.ksp.processing.SymbolProcessor
 import com.google.devtools.ksp.processing.SymbolProcessorEnvironment
 import com.google.devtools.ksp.processing.SymbolProcessorProvider
@@ -7,5 +9,10 @@ import com.google.devtools.ksp.processing.SymbolProcessorProvider
 class MockingbirdSymbolProcessorProvider : SymbolProcessorProvider {
     override fun create(
         environment: SymbolProcessorEnvironment
-    ): SymbolProcessor = MockingbirdSymbolProcessor(environment.codeGenerator, environment.logger)
+    ): SymbolProcessor = MockingbirdSymbolProcessor(
+        environment.codeGenerator,
+        environment.logger,
+        FakeImplementationGenerator(),
+        FakeFunctionGenerator()
+    )
 }
