+++ mockingbird/processor/src/main/kotlin/com/anthonycr/mockingbird/processor/internal/strings.kt
@@ -1,0 +1,6 @@
+package com.anthonycr.mockingbird.processor.internal
+
+/**
+ * Use a space character to prevent KotlinPoet from wrapping string literals.
+ */
+fun String.noWrap(): String = this.replace(" ", "·")
